improved readability of calculate_genre_match_and_calculate_avg

+++ graph.py
@@ -72,7 +72,7 @@ class ReccomenderGraph:
         self.users[friend_user].friends_list.append(self.users[user])
 
     # float is path score between 0 - 10 (actual path score avgd with the similarity)
-    def get_all_path_scores(self, depth: int, user: aau.User) -> list[tuple[aau.Anime, int]]:
+    def get_all_path_scores(self, depth: int, user: aau.User) -> list[tuple[aau.Anime, float]]:
         """Find all anime at a certain depth and calculate a path score for each anime based on
         the reviews given to it and the user's priorities, and returns the anime with the top 10 path scores
         Preconditions:
